requirements.txt
Django==1.11.3
django-guardian==1.4.9
django-markdownx==2.0.21
jsonfield==2.0.2
Markdown==2.6.8
olefile==0.44
Pillow==4.2.1
python-markdown-math==0.3
pytz==2017.2
six==1.10.0
